+++ src/Apps/Consign/Routes/MarketingLanding/Components/LandingPage/HeaderSWA.tsx
@@ -1,4 +1,3 @@
-import { ContextModule, Intent } from "@artsy/cohesion"
 import {
   Button,
   Column,
@@ -9,17 +8,13 @@ import {
   Text,
 } from "@artsy/palette"
 import { useMarketingLandingTracking } from "Apps/Consign/Routes/MarketingLanding/Utils/marketingLandingTracking"
-import { useAuthDialog } from "Components/AuthDialog"
 import { RouterLink } from "System/Components/RouterLink"
 import { useFeatureFlag } from "System/Hooks/useFeatureFlag"
-import { useSystemContext } from "System/Hooks/useSystemContext"
 import { Media } from "Utils/Responsive"
 import { resized } from "Utils/resized"
 
 export const HeaderSWA = () => {
   const getInTouchRoute = "/sell/inquiry"
-  const { isLoggedIn } = useSystemContext()
-  const { showAuthDialog } = useAuthDialog()
 
   const {
     trackStartSellingClick,
@@ -61,24 +56,6 @@ export const HeaderSWA = () => {
                   enableNewSubmissionFlow ? "sell2/intro" : "/sell/submission"
                 }
                 onClick={event => {
-                  if (!isLoggedIn) {
-                    event.preventDefault()
-
-                    showAuthDialog({
-                      mode: "Login",
-                      options: {
-                        title: () => {
-                          return "Log in to submit an artwork for sale"
-                        },
-                      },
-                      analytics: {
-                        contextModule: ContextModule.sellHeader,
-                        intent: Intent.login,
-                      },
-                    })
-
-                    return
-                  }
                   trackStartSellingClick("Header")
                 }}
                 mb={[1, 0]}
